+++ project/libs/shared/core/src/index.ts
@@ -1,8 +1,14 @@
 export { Entity } from './lib/base/entity';
 
+export { PostType } from './lib/types/post-type.enum'
+export { PostStatusType } from './lib/types/post-status-type.enum'
+
 export { User } from './lib/types/user.interface';
 export { AuthUser } from './lib/types/auth-user.interface';
-export * from './lib/types/post.interface';
+export { PostTag } from './lib/types/post-tag.interface'
+export { Comment } from './lib/types/comment.interface'
+export { PostLike } from './lib/types/post-like.interface'
+export { Post } from './lib/types/post.interface'
 
 export { StorableEntity } from './lib/interfaces/storable-entity.interface';
 export { EntityFactory } from './lib/interfaces/entity-factory.interface';
