+++ project/libs/shared/core/src/lib/types/comment.interface.ts
@@ -1,0 +1,8 @@
+export interface Comment {
+  id?: string;
+  postId?: string;
+  userId: string;
+  message: string;
+  createdAt?: Date;
+  updatedAt?: Date;
+}
