+++ project/libs/shared/core/src/lib/types/post-like.interface.ts
@@ -1,0 +1,8 @@
+
+export interface PostLike {
+  id?: string;
+  userId: string;
+  isDelete: boolean,
+  createdAt?: Date;
+  updatedAt?: Date;
+}
