+++ project/libs/shared/core/src/lib/types/post-status-type.enum.ts
@@ -1,0 +1,4 @@
+export enum PostStatusType {
+  Published = 'published',
+  Draft = 'draft'
+}
