+++ project/libs/shared/core/src/lib/types/post-tag.interface.ts
@@ -1,0 +1,6 @@
+export type PostTag = {
+  id?: string;
+  tag: string;
+  createdAt?: Date;
+  updatedAt?: Date;
+}
