+++ project/libs/shared/core/src/lib/types/post-type.enum.ts
@@ -1,0 +1,7 @@
+export enum PostType {
+  Photo = 'photo',
+  Video = 'video',
+  Text = 'text',
+  Quote = 'quote',
+  Link = 'link'
+}
